Clean

+++ SignalServiceKit/src/Loki/API/LokiAPI+SwarmAPI.swift
@@ -5,7 +5,7 @@ public extension LokiAPI {
     // MARK: Settings
     private static let minimumSnodeCount = 2 // TODO: For debugging purposes
     private static let targetSnodeCount = 3 // TODO: For debugging purposes
-    private static let defaultSnodePort: UInt32 = 8080
+    private static let defaultSnodePort: UInt16 = 8080
     
     // MARK: Caching
     private static let swarmCacheKey = "swarmCacheKey"
