+++ SignalServiceKit/src/Loki/API/TargetWrapper.swift
@@ -1,7 +1,7 @@
 
 @objc internal final class TargetWrapper : NSObject, NSCoding {
     internal let address: String
-    internal let port: UInt32
+    internal let port: UInt16
     
     internal init(from target: LokiAPI.Target) {
         address = target.address
@@ -11,7 +11,7 @@
     
     internal init?(coder: NSCoder) {
         address = coder.decodeObject(forKey: "address") as! String
-        port = coder.decodeObject(forKey: "port") as! UInt32
+        port = coder.decodeObject(forKey: "port") as! UInt16
         super.init()
     }
